Initial mapping of /sync call to local database and caching of batch identifiers

+++ composeApp/src/commonMain/kotlin/data/io/matrix/room/RoomMessageFileInfo.kt
@@ -9,7 +9,7 @@ data class RoomMessageFileInfo(
     val mimetype: String? = null,
 
     /** The size of the file in bytes. */
-    val size: Long? = null,
+    val size: Int? = null,
 
     /** Information on the encrypted thumbnail file if encrypted. */
     val thumbnailFile: MatrixEncryptedFile? = null,
